fixes made
